requirements.txt
Django==2.0.1
django-admin-tools==0.8.1
django-tinymce==2.7.0
Pillow==5.0.0
pkg-resources==0.0.0
pytz==2017.3
